+++ lib/src/b.dart
@@ -1,5 +1,6 @@
 library b;
 
+import 'dart:math';
 import 'package:characters/characters.dart';
 import 'package:invertible/invertible.dart';